% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{reorder}
\alias{reorder}
\title{Reorder the levels of factors based on the value of a second variable.}
\usage{
reorder(x, y, FUN = mean, ..., order = is.ordered(y))
}
\arguments{
\item{x}{factor}

\item{y}{a vector of the same length as x}

\item{FUN}{a function whose first argument is a vector and returns a scalar}

\item{...}{optional: extra arguments supplied to FUN}

\item{order}{logical, whether return value will be an ordered factor rather than a factor}
}
\value{
factor
}
\description{
Reorder the levels of factors based on the value of a second variable.
}
\examples{
reorder(InsectSprays$spray, InsectSprays$count)
}

