% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoofactors-package.r
\docType{package}
\name{zoofactors}
\alias{zoofactors}
\alias{zoofactors-package}
\title{zoofactors: Make factors less annoying}
\description{
Factors have driven people to extreme measures, like ordering
custom conference ribbons.
}

